backend/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
APScheduler==3.10.4
pywebpush==1.9.4
